Split "utils" and "common" modules into more specific names
