+++ test/tests/widgets/test_button.py
@@ -4,7 +4,11 @@ from pytest import fixture
 
 import toga
 
-from .common import test_background_color, test_color, test_text  # noqa: F401
+from .properties import (  # noqa: F401
+    test_background_color,
+    test_color,
+    test_text,
+)
 
 
 @fixture
